More cleanup
Magician keeps a deck now so that it can just reset it for each new display rather than create the whole deck each time.

Created a prefab for the magician card so it's easier to tweak the font size and put them all in a horizontal group for spacing.

Added another set of cards for the Assistant since the user will order the cards there. Fixed the assistant view in the flow manager as well (it was showing the magician view instead).