Update encoding order
Instead of putting the root/pivot card first, the pivot card is placed 4th, with the 3 offset encoding cards coming first. This makes it easier for the magician to know more quickly what the hidden card is after seeing the 4th card if each one is revealed slowly in order.